refactor(storage): split index_maintenance.rs into focused modules
Split the 1,994-line index_maintenance.rs into smaller, focused modules
organized by operation type:

- index_maintenance/mod.rs: Module exports
- index_maintenance/prefix.rs: Prefix truncation utilities (~35 lines)
- index_maintenance/create.rs: Index creation and drop (~520 lines)
- index_maintenance/insert.rs: Row insertion into indexes (~270 lines)
- index_maintenance/delete.rs: Row deletion from indexes (~350 lines)
- index_maintenance/update.rs: Row updates and rebuilding (~460 lines)

Benefits:
- Smaller, focused files (~300-500 lines each vs 2000 lines)
- CRUD operations clearly separated by responsibility
- Prefix truncation logic isolated for reuse
- Easier to navigate and maintain

All existing tests pass. No public API changes.

Closes #4850
